# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendora API",
        "description": "Vendora personal finance API. Use Bearer token authentication on protected endpoints.",
        "contact": {
            "email": "admin@vendora.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/financial-records": {
            "get": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "List all financial records",
                "operationId": "13b7f87c9ae697d72044de14453475a4",
                "responses": {
                    "200": {
                        "description": "Array of financial records"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Create or update a financial record",
                "operationId": "2373fd7492ca64bb5f746d388f3c6ea0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "year",
                                    "month",
                                    "monthly_income",
                                    "monthly_expenses"
                                ],
                                "properties": {
                                    "year": {
                                        "type": "integer"
                                    },
                                    "month": {
                                        "type": "integer"
                                    },
                                    "monthly_income": {
                                        "type": "number"
                                    },
                                    "monthly_expenses": {
                                        "type": "number"
                                    },
                                    "savings_goal": {
                                        "type": "number"
                                    },
                                    "savings_goal_type": {
                                        "type": "string",
                                        "enum": [
                                            "percentage",
                                            "fixed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Financial record saved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financial-records/current": {
            "get": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Get current month financial record",
                "operationId": "2de95b78080fbf9782bc8742a0648d6c",
                "responses": {
                    "200": {
                        "description": "Current month record"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financial-records/year/{year}": {
            "get": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Get financial records for a year",
                "operationId": "770cd6d61ddc4182089518fbdead5236",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of monthly records for the year"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financial-records/{financial_record}": {
            "get": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Get a single financial record",
                "operationId": "67377f736452a33552cf4e8c70aa749b",
                "parameters": [
                    {
                        "name": "financial_record",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Financial record object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Update a financial record",
                "operationId": "1fe2cc18b4d20e1be2c8603b3f159ccb",
                "parameters": [
                    {
                        "name": "financial_record",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "monthly_income": {
                                        "type": "number"
                                    },
                                    "monthly_expenses": {
                                        "type": "number"
                                    },
                                    "savings_goal": {
                                        "type": "number"
                                    },
                                    "savings_goal_type": {
                                        "type": "string",
                                        "enum": [
                                            "percentage",
                                            "fixed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Financial record updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Delete a financial record",
                "operationId": "f73fdb46ffb4e63f9aa69a14ba86238c",
                "parameters": [
                    {
                        "name": "financial_record",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Financial record deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financial-records/net-worth": {
            "get": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Get net worth overview",
                "operationId": "4bb070b958095dc0e90c81f6f3e397d2",
                "responses": {
                    "200": {
                        "description": "Net worth data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financial-records/allocation": {
            "get": {
                "tags": [
                    "Financial Records"
                ],
                "summary": "Get investment allocation breakdown",
                "operationId": "c6a423c03025d4b4ef42934211443d21",
                "parameters": [
                    {
                        "name": "account",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allocation data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}